+++ quint/src/cliCommands.ts
@@ -26,18 +26,18 @@ import {
 import { ErrorMessage } from './ErrorMessage'
 
 import { Either, left, mergeInMany, right } from '@sweet-monads/either'
-import assert, { fail } from 'assert'
+import { fail } from 'assert'
 import { EffectScheme } from './effects/base'
 import { LookupTable } from './names/base'
 import { ReplOptions, quintRepl } from './repl'
-import { OpQualifier, QuintBool, QuintEx, QuintModule } from './ir/quintIr'
+import { OpQualifier, QuintEx, QuintModule } from './ir/quintIr'
 import { TypeScheme } from './types/base'
 import { createFinders, formatError } from './errorReporter'
 import { DocumentationEntry, produceDocs, toMarkdown } from './docs'
 import { QuintError, quintErrorToString } from './quintError'
 import { TestOptions, TestResult } from './runtime/testing'
 import { IdGenerator, newIdGenerator } from './idGenerator'
-import { Outcome, SimulationTrace, SimulatorOptions } from './simulation'
+import { Outcome, SimulatorOptions, showTraceStatistics } from './simulation'
 import { ofItf, toItf } from './itf'
 import { printExecutionFrameRec, printTrace, terminalWidth } from './graphics'
 import { verbosity } from './verbosity'
@@ -515,8 +515,8 @@ function maybePrintWitnesses(verbosityLevel: number, outcome: Outcome, witnesses
 export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure<TracingStage>> {
   const simulator = { ...prev, stage: 'running' as stage }
   const startMs = Date.now()
-  // Force disable output if `--out-itf` is set
-  const verbosityLevel = prev.args.outItf ? 0 : deriveVerbosity(prev.args)
+  // Verboity level controls how much of the output is shown
+  const verbosityLevel = deriveVerbosity(prev.args)
   const mainName = guessMainModule(prev)
   const main = prev.modules.find(m => m.name === mainName)
   if (!main) {
@@ -632,52 +632,16 @@ export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure
   } else {
     // Use the typescript simulator
     const evaluator = new Evaluator(prev.resolver.table, recorder, options.rng, options.storeMetadata)
-    const evalResult = evaluator.simulate(
+    outcome = evaluator.simulate(
       init,
       step,
       invariant,
       witnesses,
       prev.args.maxSamples,
       prev.args.maxSteps,
-      prev.args.nTraces ?? 1
+      prev.args.nTraces ?? 1,
+      options.onTrace
     )
-
-    simulator.seed = recorder.bestTraces[0]?.seed
-
-    const results: Either<QuintError[], SimulationTrace[]> = mergeInMany(
-      recorder.bestTraces.map((trace, index) => {
-        const maybeEvalResult = trace.frame.result
-        if (maybeEvalResult.isLeft()) {
-          return left(maybeEvalResult.value)
-        }
-        const quintExResult = maybeEvalResult.value.toQuintEx(prev.idGen)
-        assert(quintExResult.kind === 'bool', 'invalid simulation produced non-boolean value ')
-        const simulationSucceeded = quintExResult.value
-        const status = simulationSucceeded ? 'ok' : 'violation'
-        const states = trace.frame.args.map(e => e.toQuintEx(prev.idGen))
-
-        options.onTrace(index, status, evaluator.varNames(), states)
-
-        return right({ states, result: simulationSucceeded, seed: trace.seed })
-      })
-    )
-
-    if (results.isLeft()) {
-      return cliErr('Runtime error', {
-        ...simulator,
-        errors: results.value.map(mkErrorMessage(prev.sourceMap)),
-      })
-    }
-
-    let traces = results.value
-
-    outcome = {
-      status: evalResult.isRight() ? ((evalResult.value.result as QuintBool).value ? 'ok' : 'violation') : 'error',
-      errors: evalResult.isLeft() ? [evalResult.value] : [],
-      bestTraces: traces,
-      witnessingTraces: evalResult.isRight() ? evalResult.value.witnessingTraces : [],
-      samples: evalResult.isRight() ? evalResult.value.samples : 0,
-    }
   }
 
   const elapsedMs = Date.now() - startMs
@@ -691,15 +655,20 @@ export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure
       return cliErr('Runtime error', {
         ...simulator,
         status: outcome.status,
-        // trace: states,
+        seed: prev.args.seed,
         errors: outcome.errors.map(mkErrorMessage(prev.sourceMap)),
       })
 
     case 'ok':
       maybePrintCounterExample(verbosityLevel, states, frames, prev.args.hide || [])
       if (verbosity.hasResults(verbosityLevel)) {
-        console.log(chalk.green('[ok]') + ' No violation found ' + chalk.gray(`(${elapsedMs}ms).`))
+        console.log(
+          chalk.green('[ok]') +
+            ' No violation found ' +
+            chalk.gray(`(${elapsedMs}ms at ${Math.round((1000 * outcome.samples) / elapsedMs)} traces/second).`)
+        )
         if (verbosity.hasHints(verbosityLevel)) {
+          console.log(chalk.gray(showTraceStatistics(outcome.traceStatistics)))
           console.log(chalk.gray('You may increase --max-samples and --max-steps.'))
           console.log(chalk.gray('Use --verbosity to produce more (or less) output.'))
         }
@@ -715,7 +684,11 @@ export async function runSimulator(prev: TypecheckedStage): Promise<CLIProcedure
     case 'violation':
       maybePrintCounterExample(verbosityLevel, states, frames, prev.args.hide || [])
       if (verbosity.hasResults(verbosityLevel)) {
-        console.log(chalk.red(`[violation]`) + ' Found an issue ' + chalk.gray(`(${elapsedMs}ms).`))
+        console.log(
+          chalk.red(`[violation]`) +
+            ' Found an issue ' +
+            chalk.gray(`(${elapsedMs}ms at ${Math.round((1000 * outcome.samples) / elapsedMs)} traces/second).`)
+        )
 
         // Show which invariants were violated, if we had multiple
         if (individualInvariants.length > 1) {