+++ quint/src/environment.d.ts
@@ -1,0 +1,9 @@
+declare global {
+  namespace NodeJS {
+    interface ProcessEnv {
+      QUINT_RETRY_NONDET_SMALLER_THAN?: string
+    }
+  }
+}
+
+export {}