+++ CHANGELOG.md
@@ -8,12 +8,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## UNRELEASED
 
 ### Added
+### Changed
+
+- `--out-itf` does not suppress outputs anymore. Shown output amount only depends on `--verbosity` now (#1664) 
+
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## v0.25.0 -- 2025-05-28
+
+### Added
 
 - Added `--hide` option to hide specific variables from output (#1656)
 - Errors from the Rust evaluator now show the corresponding location (#1648)
 - Added support for calling `q::debug` with a single argument, which shows both the expression and its value (#1660)
+- `quint run` now reports statistics on trace length and the simulation speec in traces/second (#1669)
 
 ### Changed
+
+- `oneOf()` now automatically handles empty sets instead of giving an error and has a retry strategy for small sets (#1670)
+
 ### Deprecated
 ### Removed
 ### Fixed